requirements.txt
discord.py>=2.0.0
aiohttp>=3.8.0
websockets>=10.0.0
